+++ changelog.md
@@ -5,6 +5,9 @@
   * Highly unstable will explode instantly
 * Added Altimeter
   * Displays roughly your altitude
+* Added "Piercing Pickaxe"
+  * Pickaxes deal bonus damage that bypasses armor
+* Removed Bone Club
 * Added Red and Brown mushroom stews
 * Out of season crops no longer break. They just don't grow
 * Florpium now generates much more but no longer drops a sh*t ton of nuggets with fortune